Remove eccentricity factor from the Hellwig2020 model
The rendering of highly saturated blue and magenta changes a
little bit.  They become slightly lighter.  In vast majority of
images there is nothing between this version and previous version
with the eccentricity factor.  This change simplifies the model.

Add comments in code how the Hellwig2020 model has been changed
in this DRT.

The custom primaries were adjusted slightly because of the change
in blue/magenta rendering.

Bump version to v29.  Remove some unused code.