Add gamut intersection approximation
Adds gamut intersection approximation to a rounded and, optionally,
smoothened triangle, replacing the old iterative gamut intersection.
For the moment the smoothing is disabled.

Removes all old tonescales and leaves the ACES2 candidate tonescale,
Daniele Evo curve.

Changes the default surround to Dim.  Rendering changes a little bit.
Chroma compression is changed as well due to this change.

Removes the focusDistanceMin and focusDistanceMax and brings back the
old single value focusDistance.  The value is set to the old 0.5 value.